feat: add continuation adapters for built-ins and containers #55
