+++ src/aleff/_multishot/v1/_aleff.c
@@ -19,7 +19,7 @@
 #include <frameobject.h>
 #include <limits.h>
 #include <stddef.h>
-#include "continuation_adapters.h"
+#include "adapters/api.h"
 #ifdef _WIN32
 #  define WIN32_LEAN_AND_MEAN
 #  include <windows.h>
