+++ src/aleff/_multishot/v1/adapters/api.h
